pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "can-blackbox"
version = "0.1.0"
description = "CAN 버스 BLF 로깅 및 AWS S3 전송 솔루션"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "CAN Blackbox Contributors"}
]
dependencies = [
    "python-can>=4.3.0",
    "awsiotsdk>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
